+++ src/main/java/org/sharetrace/experiment/ParametersExperiment.java
@@ -4,8 +4,8 @@ import java.util.Random;
 
 public final class ParametersExperiment extends SyntheticExperiment {
 
-  private double sendTolerance;
-  private double transmissionRate;
+  private float sendTolerance;
+  private float transmissionRate;
 
   public ParametersExperiment(GraphType graphType, int nNodes, int nRepeats, long seed) {
     super(graphType, seed, nRepeats);
@@ -26,20 +26,20 @@ public final class ParametersExperiment extends SyntheticExperiment {
 
   @Override
   public void run() {
-    for (transmissionRate = 0.1d; transmissionRate < 1d; transmissionRate += 0.1d) {
-      for (sendTolerance = 0.1d; sendTolerance < 1.1d; sendTolerance += 0.1d) {
+    for (transmissionRate = 0.1f; transmissionRate < 1d; transmissionRate += 0.1d) {
+      for (sendTolerance = 0.1f; sendTolerance < 1.1d; sendTolerance += 0.1d) {
         super.run();
       }
     }
   }
 
   @Override
-  protected double sendTolerance() {
+  protected float sendTolerance() {
     return sendTolerance;
   }
 
   @Override
-  protected double transmissionRate() {
+  protected float transmissionRate() {
     return transmissionRate;
   }
